remove configuration from extension to avoid comment-out guidance
This shifts to a configuration requirement rather than a configuration
recommendation for navbar and page-footer. This means that folks do not need to
modify the extension after installing.

Some formatting changes were also made, so the text has the correct indentation
when copied.

The example was updated to show how year/version can be injected using
environment variables.